Breaking up God class and general architectural cleanup
