+++ src/main/java/se/hirt/diskspace/ui/render/HitResult.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (C) 2026 Marcus Hirt
+ *
+ * This software is free:
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. The name of the author may not be used to endorse or promote products
+ *    derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESSED OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+ * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+ * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package se.hirt.diskspace.ui.render;
+
+import se.hirt.diskspace.model.DirectoryNode;
+
+/**
+ * Result of a {@link Visualization#hitTest} query. Sealed so callers can exhaustively switch on the variants without
+ * a default branch — adding a new variant becomes a compile-time prompt to revisit every consumer.
+ */
+public sealed interface HitResult permits HitResult.OnNode, HitResult.Special {
+
+	/** Cursor is over a real {@link DirectoryNode}'s painted region. */
+	record OnNode(DirectoryNode node) implements HitResult {
+	}
+
+	/**
+	 * Cursor is over one of the non-node regions a visualisation may paint: the central hub (sunburst only), the
+	 * free-space arc / cell, the unaccounted overlay, or nothing at all.
+	 */
+	enum Special implements HitResult {HUB, FREE_SPACE, UNACCOUNTED, NONE}
+}
