+++ src/main/java/se/hirt/diskspace/ui/render/NodeColorResolver.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright (C) 2026 Marcus Hirt
+ *
+ * This software is free:
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. The name of the author may not be used to endorse or promote products
+ *    derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESSED OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+ * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+ * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package se.hirt.diskspace.ui.render;
+
+import javafx.scene.paint.Color;
+import se.hirt.diskspace.model.DirectoryNode;
+
+/**
+ * Resolves the display color of a {@link DirectoryNode}. The same colour is used by both the canvas (sunburst sector
+ * / heatmap cell) and the table swatch in the right-hand pane, so centralising the lookup keeps them in lockstep.
+ *
+ * <p>Concrete implementations encapsulate palette + scheme + family-inheritance rules. The renderers don't need to
+ * know any of that — they call {@link #colorFor} and paint.</p>
+ */
+public interface NodeColorResolver {
+
+	/**
+	 * @param node the node to colour; never {@code null}
+	 * @return the fill color for {@code node}'s sector / cell. The scan root returns the scheme's surface colour
+	 * (used by the sunburst hub); file sectors get the palette's grey family; deeper descendants inherit hue
+	 * from their family root.
+	 */
+	Color colorFor(DirectoryNode node);
+}
